Do not modify default security group but attach load balancer to it.
I previously modified the default security group to allow inbound
traffic from the vpc - so that the load balancer can route traffic
to the worker nodes. But is not necessary. The better solution
(the one implemented in this commit) is to attach the default
security group to the load balancer.